+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Unreleased
 - Introduced `Texture` type and supporting infrastructure
 - Introduced `VertexArray` and `VertexBuffer` types and associated
   functionality
+- Introduced `Framebuffer` type
 
 
 0.1.0